# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "1Mao API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/chat/messages": {
            "get": {
                "description": "Retorna o histórico de mensagens entre usuários",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Buscar mensagens de chat",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID do remetente",
                        "name": "sender_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "client",
                            "professional"
                        ],
                        "type": "string",
                        "description": "Tipo do remetente",
                        "name": "sender_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 2,
                        "description": "ID do destinatário",
                        "name": "receiver_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "client",
                            "professional"
                        ],
                        "type": "string",
                        "description": "Tipo do destinatário",
                        "name": "receiver_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Message"
                            }
                        }
                    }
                }
            }
        },
        "/professional/login": {
            "post": {
                "description": "Autentica um profissional e retorna um token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login de profissional",
                "parameters": [
                    {
                        "description": "Credenciais de login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpa.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/professional/register": {
            "post": {
                "description": "Cria uma nova conta de profissional",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professionals"
                ],
                "summary": "Registrar novo profissional",
                "parameters": [
                    {
                        "description": "Dados do profissional",
                        "name": "professional",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Professional"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Professional"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/professional/{id}": {
            "get": {
                "description": "Retorna os detalhes de um profissional específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professionals"
                ],
                "summary": "Obter profissional por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do profissional",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Professional"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Profissional não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/professionals": {
            "get": {
                "description": "Retorna uma lista com todos os profissionais cadastrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Professionals"
                ],
                "summary": "Listar todos os profissionais",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Professional"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Message": {
            "description": "Estrutura completa de mensagem com metadados",
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Olá, como posso ajudar?"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "receiver_id": {
                    "type": "integer",
                    "example": 2
                },
                "receiver_type": {
                    "type": "string",
                    "enum": [
                        "client",
                        "professional"
                    ],
                    "example": "professional"
                },
                "sender_id": {
                    "type": "integer",
                    "example": 1
                },
                "sender_type": {
                    "type": "string",
                    "enum": [
                        "client",
                        "professional"
                    ],
                    "example": "client"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T15:04:05Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "domain.Professional": {
            "description": "Modelo completo de profissional",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "experience": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "httpa.LoginRequest": {
            "description": "Credenciais para login de profissional",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "professional@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "sua_senha_secreta"
                }
            }
        }
    }
}